stateId和eventId改成uint32

+++ room/ai/scxlai/dingque.go
@@ -19,17 +19,17 @@ type dingqueStateAI struct {
 func init() {
 	g := global.GetDeskAutoEventGenerator()
 	// 血流
-	g.RegisterAI(scxlGameID, majong.StateID_state_dingque, &dingqueStateAI{})
-	g.RegisterAI(scxlGameID, majong.StateID_state_huansanzhang, &huansanzhangStateAI{})
-	g.RegisterAI(scxlGameID, majong.StateID_state_chupaiwenxun, &chupaiWenxunStateAI{})
-	g.RegisterAI(scxlGameID, majong.StateID_state_zixun, &zixunStateAI{})
-	g.RegisterAI(scxlGameID, majong.StateID_state_waitqiangganghu, &waitQiangganghuStateAI{})
+	g.RegisterAI(scxlGameID, int32(majong.StateID_state_dingque), &dingqueStateAI{})
+	g.RegisterAI(scxlGameID, int32(majong.StateID_state_huansanzhang), &huansanzhangStateAI{})
+	g.RegisterAI(scxlGameID, int32(majong.StateID_state_chupaiwenxun), &chupaiWenxunStateAI{})
+	g.RegisterAI(scxlGameID, int32(majong.StateID_state_zixun), &zixunStateAI{})
+	g.RegisterAI(scxlGameID, int32(majong.StateID_state_waitqiangganghu), &waitQiangganghuStateAI{})
 	// 血战
-	g.RegisterAI(scxzGameID, majong.StateID_state_dingque, &dingqueStateAI{})
-	g.RegisterAI(scxzGameID, majong.StateID_state_huansanzhang, &huansanzhangStateAI{})
-	g.RegisterAI(scxzGameID, majong.StateID_state_chupaiwenxun, &chupaiWenxunStateAI{})
-	g.RegisterAI(scxzGameID, majong.StateID_state_zixun, &zixunStateAI{})
-	g.RegisterAI(scxzGameID, majong.StateID_state_waitqiangganghu, &waitQiangganghuStateAI{})
+	g.RegisterAI(scxzGameID, int32(majong.StateID_state_dingque), &dingqueStateAI{})
+	g.RegisterAI(scxzGameID, int32(majong.StateID_state_huansanzhang), &huansanzhangStateAI{})
+	g.RegisterAI(scxzGameID, int32(majong.StateID_state_chupaiwenxun), &chupaiWenxunStateAI{})
+	g.RegisterAI(scxzGameID, int32(majong.StateID_state_zixun), &zixunStateAI{})
+	g.RegisterAI(scxzGameID, int32(majong.StateID_state_waitqiangganghu), &waitQiangganghuStateAI{})
 
 }
 
@@ -96,7 +96,7 @@ func (h *dingqueStateAI) dingque(player *majong.Player) *interfaces.AIEvent {
 		return nil
 	}
 	return &interfaces.AIEvent{
-		ID:      majong.EventID_event_dingque_request,
+		ID:      int32(majong.EventID_event_dingque_request),
 		Context: data,
 	}
 }
